+++ apps/analog-preset-e2e/vite.config.ts
@@ -6,8 +6,9 @@ import { offsetFromRoot } from '@nx/devkit';
 // https://vitejs.dev/config/
 export default defineConfig(({ mode }) => {
   return {
-    root: 'tests',
+    root: __dirname,
     test: {
+      reporters: ['default'],
       globals: true,
       environment: 'node',
       include: ['**/*.spec.ts'],