+++ packages/nx-plugin/src/generators/app/versions/nx_18_X/versions.ts
@@ -16,13 +16,16 @@ export const V18_X_MARKED_GFM_HEADING_ID = '^3.1.0';
 export const V18_X_MARKED_HIGHLIGHT = '^2.0.1';
 export const V18_X_MERMAID = '^10.2.4';
 export const V18_X_PRISMJS = '^1.29.0';
+export const V18_X_TAILWINDCSS = '^3.0.2';
+export const V18_X_POSTCSS = '^8.4.5';
+export const V18_X_AUTOPREFIXER = '^10.4.0';
 
 // devDependencies
 export const V18_X_ANALOG_JS_PLATFORM = '^0.2.0';
 export const V18_X_NX_VITE = '^18.0.0';
 export const V18_X_NX_LINTER = '^18.0.0';
 export const V18_X_JSDOM = '^22.1.0';
-export const V18_X_VITE = '^5.0.0';
+export const V18_X_VITE = '~5.0.0';
 export const V18_X_VITE_TSCONFIG_PATHS = '^4.2.0';
 export const V18_X_VITEST = '^1.0.4';
 export const V18_X_ZOD = '^3.21.4';